pyproject.toml
[project]
name = "mia-relay"
version = "0.1.0"
description = "Mia 云转发中继：在手机与 PC Agent 之间转发 wire-protocol 信封。"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "fastapi>=0.110,<1.0",
    "uvicorn[standard]>=0.27,<1.0",
    "websockets>=12.0,<14.0",
    "pydantic>=2.6,<3.0",
    "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "httpx>=0.27,<1.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
